requirements.txt
streamlit>=1.28.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
plotly>=5.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
